refactoring : 랜딩페이지 랭킹 3위 까지로 수정, 랭킹 페이지와 폴더 합침
