+++ src/entities/rank/index.ts
@@ -1,0 +1,2 @@
+export { default as Rankings } from '@/entities/rank/ui/Rankings';
+export { getAllTimeRankings } from '@/entities/rank/actions/getRankings.actions';
