+++ src/entities/rank/types.ts
@@ -1,3 +1,10 @@
-import { IRanking } from '../rankings/actions/getRankings.actions.types';
-
 export type TAroundRanking = IRanking & { isMe: boolean };
+
+export interface IRanking {
+  userId: number;
+  nickname: string;
+  ranks: number;
+  score: number;
+}
+
+export type TRankings = IRanking[];
