+++ src/entities/rank/ui/Rankings.tsx
@@ -1,15 +1,10 @@
-import { IRanking } from '../actions/getRankings.actions.types';
+import { TRankings } from '../types';
+import NoRanking from './NoRankings';
 import RankingItem from './RankingItem';
 
 interface RankingsProps {
-  rankings: IRanking[];
+  rankings: TRankings;
 }
-
-const BADGE: Record<number, string> = {
-  1: '🏆',
-  2: '🥈',
-  3: '🥉',
-};
 
 export default function Rankings({ rankings }: RankingsProps) {
   if (!rankings) {
@@ -24,9 +19,9 @@ export default function Rankings({ rankings }: RankingsProps) {
         <div className="text-center mb-12">
           <h2 className="text-3xl font-bold mb-4">
             <span className="text-secondary">Ez-Code 종합 랭킹</span>
-            <span className="text-white">TOP 10</span>
+            <span className="text-white">TOP 3</span>
           </h2>
-          <p className="text-gray-400">Ez-Code에서 가장 활발하게 문제를 해결한 개발자들</p>
+          <p className="text-gray-400">Ez-Code에서 가장 활발하게 문제를 해결한 개발자 3인</p>
         </div>
         <div className="max-w-4xl mx-auto">
           <div className="bg-gray-900/50 border-gray-700">
@@ -36,15 +31,11 @@ export default function Rankings({ rankings }: RankingsProps) {
                   {rankings.length > 0 ? (
                     <>
                       {rankings.map((ranking) => (
-                        <RankingItem
-                          key={ranking.userId}
-                          ranking={ranking}
-                          badge={BADGE[ranking.ranks] || ''}
-                        />
+                        <RankingItem key={ranking.userId} ranking={ranking} />
                       ))}
                     </>
                   ) : (
-                    <RankingItem />
+                    <NoRanking />
                   )}
                 </ul>
               </div>
