--- src/entities/rankings/actions/getRankings.actions.types.ts
@@ -1,8 +1,0 @@
-export type TRankings = IRanking[];
-
-export interface IRanking {
-  userId: number;
-  nickname: string;
-  ranks: number;
-  score: number;
-}
